+++ src/app/core/models/dto/credential-revoke-request.dto.ts
@@ -1,4 +1,4 @@
 export interface CredentialRevokeRequestDto {
-    credentialId: string,
+    procedureId: string,
     listId: string
 }